tests/wpt/tests/html/semantics/permission-element/permission-icon/icon-color-defaults-to-text-color.html

<!DOCTYPE html>
<meta charset=utf-8>
<link rel="help" href="https://github.com/WICG/PEPC/blob/main/explainer.md#locking-the-pepc-style">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<body>

  <style>
    permission {
      color: #ff0000;
    }
  </style>

  <permission id="permission_element" type="camera"></permission>

  <script>
    promise_test(async() => {
      let permission_element = document.getElementById("permission_element");
      assert_equals(getComputedStyle(permission_element, "::permission-icon").fill, 'rgb(255, 0, 0)')
    }, "Permission element icon's color should be same as the text color unless explicitly changed.")
  </script>
</body>
